% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAdminName.R
\name{addAdminName}
\alias{addAdminName}
\title{getCountyName}
\usage{
addAdminName(
  df,
  longitude = "lon",
  latitude = "lat",
  country = "KEN",
  level = 1,
  admin.name = "NAME_1"
)
}
\arguments{
\item{df}{Dataframe with longitude and latitude columns}

\item{longitude}{Name of longitude column in df}

\item{latitude}{Name of latitude column in df}
}
\value{
dataframe with added columns
}
\description{
A function to down load data for specified administrative level and return
the dataframe with added columns
}
