% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_N_maps.R
\name{admin_sf}
\alias{admin_sf}
\title{Load/Download  shapefiles from gadm.org}
\usage{
admin_sf(country, level = 0)
}
\arguments{
\item{country}{Country code}

\item{level}{Admin level}
}
\value{
Shapefile in sf format
}
\description{
Downloads in sp format then converts to sf
}
