% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIFfunctions.R
\name{backselect}
\alias{backselect}
\title{Back select Independent variables}
\usage{
backselect(df, untouched = NULL, threshold = 10, trace = T, ...)
}
\arguments{
\item{df}{Dataframe with the Independent variables}

\item{threshold}{VIF value selection threshold}

\item{trace}{Boolean: wheter the details of each selection step should be
sent to standard output}

\item{...}{}
}
\value{

}
\description{
Backwards selection of explanatory variables, stops when all VIF values are
below the 'threshold'
}
