% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIFfunctions.R
\name{calculateVIF}
\alias{calculateVIF}
\title{VIF functions}
\usage{
calculateVIF(df, targetV = NULL, randomV = NULL, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{other arguments passed to fmsb::VIF()}

\item{cols2ignore}{a vector of column names: colnames not included when
calculating VIF values.}
}
\value{

}
\description{
A function that calculates the Variance Inflation Factor (VIF) values
using the function fmsb::VIF()and performs operations relating to VIF
}
