% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{DailytoYearyAve}
\alias{DailytoYearyAve}
\title{daily values to average daily values}
\usage{
DailytoYearyAve(list.of.df)
}
\arguments{
\item{list.of.df}{A list of dataframes. each data frame containdns the daily value}
}
\value{
a dataframe over average daily values for every day of the year
}
\description{
Calculate the average daily values for the selected years. For each day of
the year, get the mean value over the years given.
}
