% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCol.R
\name{dropCol}
\alias{dropCol}
\title{dropCol}
\usage{
dropCol(df, cols = "")
}
\arguments{
\item{df}{dataframe}

\item{cols}{names of columns to be removed from df}
}
\value{
df with column names in cols removed
}
\description{
Remove dataframe columns
}
