% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCol.R
\name{dropNaCol}
\alias{dropNaCol}
\title{dropNaCol}
\usage{
dropNaCol(df, names_only = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{names_only}{logical, whether to return dataframe or names (default is FALSE)}
}
\value{
df with NA values of names of columns with Na values
}
\description{
Remove dataframe columns with Na values
}
