% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractraster.R
\name{extractadmin}
\alias{extractadmin}
\title{Extract mean value per area}
\usage{
extractadmin(folderpath, country = "KEN", level = 1, years = 1981:2010)
}
\arguments{
\item{folderpath}{Path to folder with '*.nc' files}

\item{country}{abbreviation for country name (see https://gadm.org)}

\item{level}{Administration level}

\item{years}{A vector of years to select}
}
\value{
Returns a list of dataframes for each year selected
}
\description{
Extract mean values for the selected admin level
}
\details{
This function extracts data from ncd4(.nc) files in the given folder, and
returns the mean for every administration unit in the given administration
level. Not all files in the folder are used(extracted from). The function
selects file with the given years in their file name eg for 2010 you can
have precip.2008.nc, 2008.nc, Tmin008.nc etc
}
