% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractraster.R
\name{extractYearlyMean}
\alias{extractYearlyMean}
\title{Admin level mean per year}
\usage{
extractYearlyMean(
  folderpath,
  country = "KEN",
  level = 1,
  years = 1981:2010,
  extractnewdata = FALSE,
  extracted.data = NULL
)
}
\arguments{
\item{folderpath}{Path to folder with '*.nc' files}

\item{country}{abbreviation for country name (see https://gadm.org)}

\item{level}{Administration level}

\item{years}{A vector of years to select}

\item{extractnewdata}{}
}
\value{

}
\description{
Get the mean value in the admin level in the selected years
}
