% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorby.R
\name{factorby}
\alias{factorby}
\title{Factor by reference value}
\usage{
factorby(df, cols, refvalues)
}
\arguments{
\item{df}{object of class data.frame}

\item{cols}{vector of column names to be converted to factors}

\item{refvalue}{vector of reference values: same length as cols}
}
\value{
data.frame with specified columns factored
}
\description{
Convert column to factors with the specified value as the reference value
}
