% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorby.R
\name{factorby_occur}
\alias{factorby_occur}
\title{Factor with mode as reference value}
\usage{
factorby_occur(df, cols)
}
\arguments{
\item{df}{object of class data.frame}

\item{cols}{vector of column names to be converted to factors}
}
\value{
data.frame with specified columns factored
}
\description{
Convert column to factors with the mode value as the reference value
}
