% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fixedEffetsModel}
\alias{fixedEffetsModel}
\title{Fixed Effects model}
\usage{
fixedEffetsModel(df, targetv, fixedv = NULL)
}
\arguments{
\item{df}{dataframe with target variable and fixed variables}

\item{fixedv}{A vector of fixed variables names. If null all the other
columns in the dataframe}

\item{target}{The name of the target variable column}
}
\value{
Fixed effects model
}
\description{
Fixed Effects model
}
