% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIFfunctions.R
\name{forwardselect}
\alias{forwardselect}
\title{Forward select Independent variables}
\usage{
forwardselect(
  df,
  targetV,
  randomV = NULL,
  curr.model.cols,
  cols.to.add,
  threshold = 5,
  trace = FALSE
)
}
\arguments{
\item{df}{Dataframe with the Indipendent variables}

\item{curr.model.cols}{Variables that must be in the model}

\item{cols.to.add}{Variables to be forward selected}

\item{threshold}{The forward selection threshold}

\item{trace}{Boolean: wheter the details of each selection step should be
sent to standard output}
}
\value{
a list
}
\description{
Forward selection of Independent variables, starts with variables specified
by the user and incrementally adds other columns to the list while checking
the VIF scores at each step. The process stops when addition of any of the
remaining variables will yield a Vif score greater than the threshold.
}
