% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Select an optimal model}
\usage{
model_selection(
  step.direction,
  df,
  targetV,
  randomV,
  fixedv_1 = NULL,
  fixedv_2,
  trace_ = FALSE
)
}
\arguments{
\item{step.direction}{The direction of model iteration. Must be either
'forward' or 'backward'}

\item{df}{a dataframe with columns of the target variable, random variables
and fixed (fixedv_1, fixedv_2  )variables}

\item{targetV}{The target variable}

\item{randomV}{A vector of names of random variables columns in df}

\item{fixedv_1}{A vector of names of Independent variables columns in df, which
must be in the final model. If a name appears in both fixedv_1 and fixedv_2,
the fixedv_1 entry is ignored.}

\item{fixedv_2}{A vector of names of Independent variables columns in df, which
can be in the final model based on the model's accuracy. If a name appears
in both fixedv_1 and fixedv_2, the fixedv_1 entry is ignored.}

\item{trace_}{Boolean: wheter the details of each selection step should be
sent to standard output.}
}
\value{
A object of type list containing the final model
}
\description{
Function to select a model by interactively adding or removing Independent
variables to/from model, while calculating the effectsize of each model. The
function returns the model with the smallest effectsize.
}
