% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection_1}
\alias{model_selection_1}
\title{Forward Select or Back select an optimal model}
\usage{
model_selection_1(
  df,
  targetV,
  randomV,
  fixedv_1,
  fixedv_2,
  Num.of.conbinations,
  trace_ = FALSE
)
}
\arguments{
\item{df}{dataframe}

\item{targetV}{target variable}

\item{randomV}{a character vector of random variables}

\item{Num.of.conbinations}{a vector of order of combinations}
}
\description{
Function to select a model by interactively adding or removing the number of
Independent variables to the model, while calculating the effectsize of
each model. The function returns the model with the smallest effectsize.
}
