% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection_pvalue}
\alias{model_selection_pvalue}
\title{Select a model based on p-values}
\usage{
model_selection_pvalue(df, targetV, randomV, fixedv = NULL, threshold_ = 0.05)
}
\arguments{
\item{df}{a dataframe with columns of the target variable, random variables
and fixed (fixedv_1, fixedv_2  )variables}

\item{targetV}{The target variable}

\item{randomV}{A vector of names of random variables columns in df}
}
\value{
Model selected based on the p-value of fixed variables
}
\description{
random values are not used
}
