% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{normalization}
\usage{
normalization(df, cols, method = "standardization")
}
\arguments{
\item{df}{dataframe}

\item{cols}{vector of df columns to be normalized, must be one of
'standardization', 'centering', 'unit-range', 'squares',
'clipping' or 'logarithmic'}

\item{method}{Normalization method}
}
\value{
df with specified columns normalized
}
\description{
Normalize the columns
}
\details{
Will normalize the  specified columns in the df via the specified method
}
