% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{normalization_warn}
\alias{normalization_warn}
\title{Alerts to the user}
\usage{
normalization_warn(method, col_)
}
\arguments{
\item{method}{Normalization method}

\item{col_}{Column with the error}
}
\description{
Create warning for normalization function
}
\details{
Given the name of a method an a column, this function will create a warmning
message stating that the column has not been normalized since it is non-numeric.
}
