% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{organizedf}
\alias{organizedf}
\title{Organised dataframe columns by given order}
\usage{
organizedf(df, col_ordr = NULL)
}
\arguments{
\item{df}{dataframe}

\item{col_ordr}{column order}
}
\value{
list of input and output dataframe
}
\description{
Organised dataframe columns by given order
}
