% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{progressReport}
\alias{progressReport}
\title{Progress report}
\usage{
progressReport(i, total_i, starttime, Message_ = "Progress Report:")
}
\arguments{
\item{i}{Current count of items i}

\item{total_i}{The number of items i}

\item{starttime}{The time the process started}

\item{Message_}{Short message describing the process}
}
\description{
Show progress of a process
}
