% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{pvalues}
\alias{pvalues}
\title{P-Values}
\usage{
pvalues(df, targetv, fixedv = NULL)
}
\arguments{
\item{df}{dataframe with target variable and fixed variables}

\item{fixedv}{A vector of fixed variables names. If null all the other
columns in the dataframe}

\item{target}{The name of the target variable column}
}
\value{
Summary linear model with p_values
}
\description{
P-Values
}
