% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesCreate.R
\name{table_3}
\alias{table_3}
\title{Create caterpillar plot}
\usage{
table_3(modeldata = NULL)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Create caterpillar plot that shows the model variable plus the standard coefficients range
}
