% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmodel.R
\name{testmodel}
\alias{testmodel}
\title{Make predictions}
\usage{
testmodel()
}
\value{
dataframe with computed and observed values
}
\description{
Use the moddel from the main function to make predictions
}
