% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldModel.R
\name{yieldModel}
\alias{yieldModel}
\title{Run yield model}
\usage{
yieldModel(df, targetv, randomv, fixedv)
}
\arguments{
\item{df}{dataframe}

\item{targetv}{target variable}

\item{randomv}{a character vector of random variables}

\item{fixedv}{a character vector of fixedd variables}
}
\value{
model object
}
\description{
The function creates and runs a yield estimate model.
}
\details{
The function can be used to create a yield estimate model by specifying the
model variable as either target variable, random variables or fixed variables
}
\seealso{
Other Yield estimate functions: 
\code{\link{yieldModel2}()}
}
\concept{Yield estimate functions}
