% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldModel.R
\name{yieldModel2}
\alias{yieldModel2}
\title{Run yield model}
\usage{
yieldModel2(df, targetv, randomv, fixedv, use.rest.as.fixed = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{targetv}{target variable}

\item{randomv}{a character vector of random variables}

\item{fixedv}{a character vector of fixedd variables}

\item{use.rest.as.fixed}{boolean: All other variables should be fixed variables
default is TRUE (Overrides fixedv)}
}
\value{
model object
}
\description{
wrapper function for yieldModel.
}
\details{
This function takes in the target variable and the fixed variables and
considers the remaining variables to be fixed variables
}
\seealso{
Other Yield estimate functions: 
\code{\link{yieldModel}()}
}
\concept{Yield estimate functions}
