% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldModel.R
\name{yieldModel_lm}
\alias{yieldModel_lm}
\title{Run fixed effects yield model}
\usage{
yieldModel_lm(df, targetv, fixedv = NULL, use.rest.as.fixedv = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{targetv}{target variable}

\item{fixedv}{a character vector of fixed variables. Ignored if
use.rest.as.fixed is TRUE}

\item{use.rest.as.fixedvboolean:}{All other variables should be fixed variables
default is TRUE (Overrides fixedv)}
}
\value{
model object
}
\description{
The function creates and runs a fixed effects yield estimate model.
}
\details{
The function will create a yield estimate model by specifying the
target variable and fixed variables
}
