% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIFfunctions.R
\name{backselect}
\alias{backselect}
\title{Title}
\usage{
backselect(df, threshold = 10, trace = T, ...)
}
\arguments{
\item{df}{dataframe}

\item{threshold}{VIF threshold}

\item{trace}{trace}

\item{...}{}
}
\value{

}
\description{
Backwards selection of explanatory variables, stops when all VIF values are
below the 'threshold'
}
