% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{normalization}
\usage{
normalization(
  df,
  cols,
  method = "standardization",
  group = FALSE,
  groupby = "aez"
)
}
\arguments{
\item{df}{dataframe}

\item{cols}{vector of df columns to be normalized, must be one of
'standardization', 'centering', 'unit-range', 'squares',
'clipping' or 'logarithmic'}

\item{method}{Normalization method}

\item{group}{Logical: whether to normalize in groups (Works only with standardization)}

\item{groupby}{Column to be used to group. Ignored if group is false}
}
\value{
df with specified columns normalized
}
\description{
Normalize the columns
}
\details{
Will normalize the  specified columns in the df via the specified method
}
