% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldModel.R
\name{yieldModel}
\alias{yieldModel}
\title{Run mixed effects yield model}
\usage{
yieldModel(df, targetV, randomV, fixedV = NULL, use.rest.as.fixedv = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{targetV}{target variable}

\item{randomV}{a character vector of random variables}

\item{fixedV}{a character vector of fixed variables. Ignored if
use.rest.as.fixed is TRUE}

\item{use.rest.as.fixed}{boolean: All other variables should be fixed variables
default is FALSE (Overrides fixedV)}
}
\value{
model object
}
\description{
The function creates and runs a yield estimate model.
}
\details{
The function can be used to create a yield estimate model by specifying the
model variable as either target variable, random variables or fixed variables
}
\concept{Yield estimate functions}
